+++ lib/sidekiq/client.rb
@@ -1,7 +1,6 @@
 require 'multi_json'
 require 'redis'
 
-require 'sidekiq/redis_connection'
 require 'sidekiq/middleware/chain'
 require 'sidekiq/middleware/client/unique_jobs'
 
@@ -19,11 +18,11 @@ module Sidekiq
     end
 
     def self.registered_workers
-      Sidekiq.redis.smembers('workers')
+      Sidekiq.redis { |x| x.smembers('workers') }
     end
 
     def self.registered_queues
-      Sidekiq.redis.smembers('queues')
+      Sidekiq.redis { |x| x.smembers('queues') }
     end
 
     def self.queue_mappings
@@ -43,7 +42,7 @@ module Sidekiq
       pushed = false
       Sidekiq.client_middleware.invoke(item, queue) do
         payload = MultiJson.encode(item)
-        Sidekiq.redis.with_connection do |conn|
+        Sidekiq.redis do |conn|
           conn.multi do
             conn.sadd('queues', queue)
             conn.rpush("queue:#{queue}", payload)