+++ lib/sidekiq/middleware/server/unique_jobs.rb
@@ -5,7 +5,7 @@ module Sidekiq
         def call(*args)
           yield
         ensure
-          Sidekiq.redis.del(Digest::MD5.hexdigest(MultiJson.encode(args[1])))
+          Sidekiq.redis {|conn| conn.del(Digest::MD5.hexdigest(MultiJson.encode(args[1]))) }
         end
       end
     end